+++ src/main/java/com/inlym/lifehelper/requestlog/service/RequestLogService.java
@@ -44,7 +44,7 @@ public class RequestLogService {
         if (!whiteList.contains(path)) {
             CustomRequestContext context = (CustomRequestContext) request.getAttribute(CustomRequestContext.NAME);
 
-            int userId = context.getUserId() == null ? 0 : context.getUserId();
+            long userId = context.getUserId() == null ? 0 : context.getUserId();
             String ip = context.getClientIp();
             String requestId = context.getTraceId();
