+++ src/main/java/com/inlym/lifehelper/user/account/entity/User.java
@@ -28,7 +28,7 @@ import java.time.LocalDateTime;
 public class User {
     /** 用户 ID */
     @Id(keyType = KeyType.Auto)
-    private Integer id;
+    private Long id;
 
     /** 微信小程序用户唯一标识 */
     private String openid;
@@ -39,7 +39,7 @@ public class User {
      * <h2>主要用途
      * <p>给用户展示的 ID。
      */
-    private Integer accountId;
+    private Long accountId;
 
     /** 注册时间 */
     private LocalDateTime registerTime;
